zeromq: Fixup do_install_ptest_base failure
The following error is being hit:

| install: cannot stat
|  '.../tmp/work/core2-64-poky-linux/zeromq/4.2.3-r0/build/.libs/test_*':
|  No such file or directory

This is being hit in the 'state of the world' builds (see [oe] State
of the world, 2018-03-19). The uprev to v4.2.3 (commit 50eb3504c83c)
did not take into account that the location of the tests changed in
the 'build' directory as part of the new release. Updating to use the
proper directory.

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>
Signed-off-by: Khem Raj <raj.khem@gmail.com>
Signed-off-by: Armin Kuster <akuster808@gmail.com>